backend/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
faiss-cpu
unstructured
python-dotenv
validators
nltk
rouge-score
yt-dlp
tiktoken
langchain-text-splitters
rank_bm25
youtube-transcript-api
fastapi
uvicorn
dill 
mlflow
langchain-groq
ollama
ragas
datasets
python-multipart
